now displaying version number
