+++ src/components/WorkspaceContainer.js
@@ -1,7 +1,6 @@
 import {
   useContext,
   useEffect,
-  useMemo,
   useState,
 } from 'react'
 import * as isEqual from 'deep-equal'
@@ -20,7 +19,7 @@ import {
   OT_ORIG_LANG_BIBLE,
 } from 'single-scripture-rcl'
 import ResourceCard from '@components/ResourceCard'
-import { getBuildId, getResourceBibles } from '@utils/resources'
+import { getResourceBibles } from '@utils/resources'
 import { StoreContext } from '@context/StoreContext'
 import { NT_BOOKS } from '@common/BooksOfTheBible'
 import useLocalStorage from '@hooks/useLocalStorage'
@@ -64,8 +63,6 @@ function WorkspaceContainer() {
       setCurrentLayout,
     },
   } = useContext(StoreContext)
-
-  const buildId = useMemo(getBuildId, [])
 
   const layout = {
     widths: [
